Optimize disclaimer formating
