+++ .github/workflows/publish.yml
@@ -52,8 +52,8 @@ jobs:
         uses: ./.github/actions/get-version
       - name: Install dependencies
         run: npm i
-      - name: Reformat README to regular markdown
-        run: node ./scripts/release/reformat-readme.js "${{ steps.get-version.outputs.version }}"
+      - name: Reformat to regular markdown
+        run: node ./scripts/release/reformat-markdown.js "${{ steps.get-version.outputs.version }}"
       - name: Update SDK Typescript definitions
         run: node ./scripts/release/update-sdk-ts-defs.js
       - name: Build NodeJS package
@@ -74,8 +74,8 @@ jobs:
         uses: kceb/git-message-action@v2
       - id: get-version
         uses: ./.github/actions/get-version
-      - name: Replace CDN theme image links from README
-        run: node ./scripts/release/reformat-readme.js "${{ steps.get-version.outputs.version }}"
+      - name: Reformat to regular markdown
+        run: node ./scripts/release/reformat-markdown.js "${{ steps.get-version.outputs.version }}"
       - name: Configure GIT credentials
         run: |
           git config user.name "${GITHUB_ACTOR}"
